config: Moving high level functions from config_file to utils
Some high level functions are useful in many place in the code so let's
move them into utils to keep config_file only having the basic internals to manipulate the config files & defaults.
